asset-management/pyproject.toml
[tool.poetry]
name = "Asset Management"
version = "0.1.0"
description = "Asset Management for eGA-RIDC"
authors = ["Asset Team"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.4"
qrcode = "^6.1"
djangorestframework = "^3.12.4"
djangorestframework-jwt = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
